fix(auth): keep OIDC startup warnings off the optional crypto path (#3825)
